feat: refactor to module function
